README.proxy.md: update to show how to use requests on localhost
If a proxy is being used, the original socket.socket module will have
been overridden by pysocks to enable proxy support.  If a callback wants
to make a requests call to the localhost, then it will not work because
requests will try to go over the proxy.  The work around is to restore
the socket.socket module before instantiating requests.  See example in
this commit.

Signed-off-by: Paul Barrette <paulbarrette@gmail.com>